split GhciTop.hs into InteractionTop.hs and GhciTop.hs
The motivation is that InteractionTop.hs is a pure module
which contains most of the information, and the new GhciTop.hs
is a relatively small module which imports InteractionTop.hs
and which contains the impure part (the global 'theState'
mutable variable created by unsafePerformIO).

A proper interactive frontend can import the pure InteractionTop.hs
only so it will have a global state, so it could serve
multiple Agda source file editing at the same time.