run interaction tests with agdaghci
Interaction tests with ".in" extension
are interpreted with agdaghci;
interaction tests with ".in_ghci" extension
are interpreted with ghci.

agdaghci interprets interaction tests faster.
Also we can test agdaghci in this way.

Currenty only one test case is interpreted
with ghci which cannot be interpreted with
agdaghci at the moment.